api/requirements/common.txt
Django~=3.1.1
djangorestframework
pillow
django-cors-headers